requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
sqlalchemy==2.0.48
psycopg2-binary==2.9.11
alembic==1.18.4
pydantic==2.12.5
pydantic-settings==2.13.1
python-dotenv==1.2.2
httpx==0.28.1
python-jose[cryptography]==3.3.0